Udemy - Chunk Movement
Here we added some texture to the Chunk prefab, cleaned up the shadow of the player by adjusting the normal bias value to 0 in PC_RP Asset and cleaned the code to include a list of instances of chunks and organized it overall. Also changed the code from last time into a funcion instead.