Udemy - C# Lists
We are using lists and not arrays because lists are more flexible. While their speed may vary they do not have a fixed size to the elements inside it unlike an array. We refactored the code from before and included a new part to the code called SpawnChunk() and Destroy(chunk) to remove from the list and instantiate new chunks in front of the player.