Udemy - Physics Materials
Here we changed the obstacle prefab to a cart. I had to add two box colliders to cover it's entire frame and to prevent it from getting stuck on the chunks I created a Physics Material and adjusted the values to make it preferable, like adding maximum bounciness and less static.