Udemy - Physics Collision Setup
Here we added rigidbody to the fences and capsule colliders to the pickups. We started using tags, firstly with the Player character and new defined in our new code Pickup the tag using a const so that the String value can not change for any reason. Here we just set it up and used debug logs to see that the player was interacting with the fences, pickups and obstacles.