Udemy - Destroy Obstacles
Here we just added a script and massive empty game object that will destroy any object that touches it. We do this because we are instantiating a lot of obstacles on the scene and they were not being destroyed when they fell out of view, so this just ensures that unneeded elements get destroyed and the frame rate stays consistent.