Udemy - Modifying Physics
Apples make you faster but hitting an obstacle or fence will slow you down. Now a big challenge here was that if the player moved too fast the obstacle would stop falling towards you in a natural way, which we fixed by simply editing the Physics engine through our code and editing only on the z-axis.