Udemy - Cinemachine Camera Controller
Here we download the Cinemachine package from package manager and created a camera controller script to allow us to create a dynamic, cool FoV when we are speeding up and slowing down. To prevent it from getting out of control I added a minimum and maximum FoV and to prevent the FoV from decreasing or increasing at the same time I specified to StopAllCoroutines() so that it does whicher animation first.