Udemy - Speed Up Particle System
Here we added a particle effect to our cinemachine camera so that there is a cartoony effect to when the player speeds up by touching an apple. Now, you can added a particle effect by just right clicking on the cinemachine camera and selecting effects, and after some fine-tuning, making a custom made material for the effect and accounting for it in the camera controller the deed was done.