Udemy - Dependency Injection
Dependency injection means that program receives the things it needs from the outside instead of making it itself. So here we are just having the Apple and Coin rely on the various managers to instantiate them instead of what was done before, the two pickups instantiating themselves. Like Coin taking the ScoreManager as a dependency to instantiate it.