Udemy - Time Lose Condition
Here we finally create a GameManager script to manage the time functionalities in the game and maybe later on control things like restarting. When the time runs up a big GAME OVER will show up on the screen, time will slow down and the player can not control the character anymore.