Udemy - Cinemachine Composer
Here we just added some camera movement to the player when they move from side to side so it does not rigidly focus on the same spot. Mind you, I enabled a deadzone to ensure that the jump from side to side is not big either, just enough to feel flexible.